requirements.txt
garminconnect==0.1.18
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
pandas==2.0.0
python-dotenv==1.0.0 